HeThongTiemChung/HeThongTiemChung/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.1
Django==5.1.6
django-allauth==65.7.0
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-extensions==4.0
django-js-asset==3.0.1
django-oauth-toolkit==3.0.1
django-rest-passwordreset==1.5.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-oauth==1.1.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kombu==5.5.3
MarkupSafe==3.0.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
